+++ opamp/opentelemetry-opamp-client/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Breaking change: callback class `Callbacks` renamed to `OpAMPCallbacks`
+  ([#4355](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/4355))
+
 ## Version 0.1b0 (2026-03-23)
 
 - Initial implementation